+++ sshd-core/src/test/java/org/apache/sshd/common/forward/AbstractServerCloseTestSupport.java
@@ -29,12 +29,10 @@ import java.nio.channels.CompletionHandler;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 
-import org.apache.sshd.common.io.DefaultIoServiceFactoryFactory;
 import org.apache.sshd.common.util.net.SshdSocketAddress;
 import org.apache.sshd.util.test.BaseTestSupport;
 import org.junit.After;
 import org.junit.Assert;
-import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -147,7 +145,7 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
         outputDebugMessage("readInOneBuffer(port=%d)", serverPort);
         outputDebugMessage("expecting %d bytes", PAYLOAD.length());
         try (Socket s = new Socket()) {
-            s.setSoTimeout(300);
+            s.setSoTimeout(2000);
             s.setReceiveBufferSize(65536);
             s.connect(new InetSocketAddress(TEST_LOCALHOST, serverPort));
             Thread.sleep(50L);
@@ -167,7 +165,7 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
         outputDebugMessage("readInTwoBuffersWithPause(port=%d)", serverPort);
         outputDebugMessage("expecting %d bytes", PAYLOAD.length());
         try (Socket s = new Socket()) {
-            s.setSoTimeout(300);
+            s.setSoTimeout(2000);
             s.setReceiveBufferSize(65536);
             s.connect(new InetSocketAddress(TEST_LOCALHOST, serverPort));
             Thread.sleep(50L);
@@ -211,19 +209,6 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
         return true;
     }
 
-    // The tests that attempt to read the payload in a single or in two buffers are fundamentally flawed; they can only
-    // work if all transport buffers are large enough to hold the full or half payload. This appears to be the case with
-    // NIO2 and Netty, but MINA manages the read buffer size dynamically, starting out with 2048 bytes. This is too
-    // small for the payload.
-    //
-    // In general these tests are flawed because one should _never_ read an expected number of bytes in a single read
-    // and assume one has got it all. Data may be broken up at any layer. The correct way to read an expected number of
-    // bytes is _always_ a read loop, breaking when read returns -1 (not zero).
-    protected boolean isMina() {
-        return "MinaServiceFactoryFactory".equals(DefaultIoServiceFactoryFactory.getDefaultIoServiceFactoryFactoryInstance()
-                .getIoServiceProvider().getClass().getSimpleName());
-    }
-
     /*
      * Connect to test server via port forward and read real quick with one big buffer.
      *
@@ -231,7 +216,6 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
      */
     @Test
     public void testRemotePortForwardOneBuffer() throws Exception {
-        Assume.assumeTrue("This test cannot work on MINA", !isMina());
         SshdSocketAddress pf = startRemotePF();
         try {
             readInOneBuffer(pf.getPort());
@@ -247,7 +231,6 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
      */
     @Test
     public void testRemotePortForwardTwoBuffers() throws Exception {
-        Assume.assumeTrue("This test cannot work on MINA", !isMina());
         SshdSocketAddress pf = startRemotePF();
         try {
             readInTwoBuffersWithPause(pf.getPort());
@@ -268,7 +251,6 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
 
     @Test
     public void testLocalPortForwardOneBuffer() throws Exception {
-        Assume.assumeTrue("This test cannot work on MINA", !isMina());
         SshdSocketAddress pf = startLocalPF();
         try {
             readInOneBuffer(pf.getPort());
@@ -284,7 +266,6 @@ public abstract class AbstractServerCloseTestSupport extends BaseTestSupport {
      */
     @Test
     public void testLocalPortForwardTwoBuffers() throws Exception {
-        Assume.assumeTrue("This test cannot work on MINA", !isMina());
         SshdSocketAddress pf = startLocalPF();
         try {
             readInTwoBuffersWithPause(pf.getPort());